Decouple AgentLoader from method tracer
AgentLoader is now serving many different tracers, so it makes sense to
decouple this class from the method tracer. This organizes the project
into a small and tidy tree of package dependencies.
