+++ src/main/java/com/google/idea/perf/methodtracer/ArgSet.kt
@@ -16,8 +16,6 @@
 
 package com.google.idea.perf.methodtracer
 
-import com.google.idea.perf.agent.Argument
-
 class ArgSet(val items: Array<Argument>) {
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
